requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0
tensorflow>=2.13.0
requests>=2.31.0
torch>=2.0.0
